chore: bump greptime-proto to replace region_dir
