+++ src/common/base/src/lib.rs
@@ -15,7 +15,6 @@
 pub mod bit_vec;
 pub mod buffer;
 pub mod bytes;
-pub mod paths;
 #[allow(clippy::all)]
 pub mod readable_size;
