+++ src/file-table-engine/src/test_util.rs
@@ -20,7 +20,8 @@ use datatypes::prelude::ConcreteDataType;
 use datatypes::schema::{ColumnSchema, RawSchema, Schema, SchemaBuilder, SchemaRef};
 use object_store::services::Fs;
 use object_store::ObjectStore;
-use table::engine::{table_dir, EngineContext, TableEngine};
+use store_api::path_utils::table_dir;
+use table::engine::{EngineContext, TableEngine};
 use table::metadata::{RawTableInfo, TableInfo, TableInfoBuilder, TableMetaBuilder, TableType};
 use table::requests::{self, CreateTableRequest, TableOptions};
 use table::TableRef;
