+++ src/store-api/src/lib.rs
@@ -19,6 +19,7 @@ pub mod data_source;
 pub mod logstore;
 pub mod manifest;
 pub mod metadata;
+pub mod path_utils;
 pub mod region_engine;
 pub mod region_request;
 pub mod storage;
