+++ src/table/src/engine.rs
@@ -15,10 +15,9 @@
 use std::fmt::{self, Display};
 use std::sync::Arc;
 
-use common_base::paths::DATA_DIR;
 use common_procedure::BoxedProcedure;
 use datafusion_common::TableReference as DfTableReference;
-use store_api::storage::{RegionId, RegionNumber};
+use store_api::storage::RegionNumber;
 
 use crate::error::{self, Result};
 use crate::metadata::TableId;
@@ -187,25 +186,6 @@ pub trait TableEngineProcedure: Send + Sync {
 
 pub type TableEngineProcedureRef = Arc<dyn TableEngineProcedure>;
 
-/// Generate region name in the form of "{TABLE_ID}_{REGION_NUMBER}"
-#[inline]
-pub fn region_name(table_id: TableId, region_number: RegionNumber) -> String {
-    format!("{table_id}_{region_number:010}")
-}
-
-#[inline]
-pub fn table_dir(catalog_name: &str, schema_name: &str, table_id: TableId) -> String {
-    format!("{DATA_DIR}{catalog_name}/{schema_name}/{table_id}/")
-}
-
-pub fn region_dir(catalog_name: &str, schema_name: &str, region_id: RegionId) -> String {
-    format!(
-        "{}{}",
-        table_dir(catalog_name, schema_name, region_id.table_id()),
-        region_name(region_id.table_id(), region_id.region_number())
-    )
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -219,14 +199,5 @@ mod tests {
         };
 
         assert_eq!("greptime.public.test", table_ref.to_string());
-    }
-
-    #[test]
-    fn test_region_dir() {
-        let region_id = RegionId::new(42, 1);
-        assert_eq!(
-            region_dir("my_catalog", "my_schema", region_id),
-            "data/my_catalog/my_schema/42/42_0000000001"
-        );
     }
 }
